Each view has its own file now. Some PEP8 and removed subtransaction.
