+++ src/senaite/sync/browser/configure.zcml
@@ -12,14 +12,14 @@
   <browser:page
       for="Products.CMFPlone.interfaces.IPloneSiteRoot"
       name="do_auto_sync"
-      class=".views.AutoSync"
+      class=".autosync.AutoSync"
       permission="cmf.ManagePortal"
       />
 
   <browser:page
       for="Products.CMFPlone.interfaces.IPloneSiteRoot"
       name="edit_auto_sync"
-      class=".views.EditAutoSync"
+      class=".autosync.EditAutoSync"
       permission="cmf.ManagePortal"
       />
